[step4] 정적 페이지 요청 처리 클래스 별도 분리, Reflection한 메서드들 실행 시 내부에서 발생하는 CustomException ResponseHandler까지 전달되도록 구현
