+++ src/main/java/webserver/route/Route.java
@@ -2,39 +2,25 @@ package webserver.route;
 
 import exception.CustomErrorType;
 import exception.CustomException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
 import webserver.route.requestMapping.RequestMapping;
 import webserver.route.requestMapping.RequestMappingFinder;
+import webserver.route.staticPage.StaticPageRequestManager;
 import webserver.route.user.requestManager.UserRequestManager;
 import webserver.type.HttpMethod;
-import webserver.type.MIMEType;
-import webserver.utils.PropertyUtils;
 import webserver.utils.URLUtils;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static webserver.utils.PropertyUtils.loadStaticSourcePathFromProperties;
-
 public class Route {
-    private static final String STATIC_SOURCE_PATH = loadStaticSourcePathFromProperties();
-    private static final String DEFAULT_HTML = PropertyUtils.loadProperties().getProperty("defaultHtml");
-    private static final String ERROR_404_HTML = PropertyUtils.loadProperties().getProperty("error404Html");
-
-    private final Map<String, String> STATIC_MAPPING = Map.of(
-            "/", DEFAULT_HTML,
-            "/main", DEFAULT_HTML,
-            "/article", DEFAULT_HTML,
-            "/login", DEFAULT_HTML,
-            "/registration", DEFAULT_HTML,
-            "default", ERROR_404_HTML
-    );
+    private static final Logger logger = LoggerFactory.getLogger(Route.class);
 
     private static final List<Class<?>> RequestManagerClasses =
             List.of(UserRequestManager.class);
@@ -43,8 +29,11 @@ public class Route {
 
     private static final Route instance = new Route();
 
+    private final StaticPageRequestManager staticPageRequestManager;
+
     private Route() {
         init();
+        staticPageRequestManager = new StaticPageRequestManager();
     }
 
     public static Route getInstance() {
@@ -80,7 +69,7 @@ public class Route {
         }
     }
 
-    public HttpResponse route(HttpRequest httpRequest) throws Exception{
+    public HttpResponse route(HttpRequest httpRequest){
         HttpMethod httpMethod = httpRequest.getHttpMethod();
         String path = URLUtils.getPath(httpRequest.getURL());
 
@@ -89,62 +78,33 @@ public class Route {
 
         // 매핑 정보가 있으면
         if (requestMappingFinder != null) {
-            Class<?> clazz = requestMappingFinder.getClazz();
-            Object o = clazz.getDeclaredConstructor().newInstance();
+            try {
+                Class<?> clazz = requestMappingFinder.getClazz();
+                Object o = clazz.getDeclaredConstructor().newInstance();
 
-            Method methodInfo = requestMappingFinder.getMethod(path);
-            HttpMethod httpMethodInfo = requestMappingFinder.getHttpMethod(path);
+                Method methodInfo = requestMappingFinder.getMethod(path);
+                HttpMethod httpMethodInfo = requestMappingFinder.getHttpMethod(path);
 
-            // 매핑된 메서드가 있고, 요청의 httpMethod가 일치하면 해당 메서드 실행(httpRequest 전달)
-            if (methodInfo != null && httpMethod.equals(httpMethodInfo)) {
-                return (HttpResponse) methodInfo.invoke(o, httpRequest);
+                // 매핑된 메서드가 있고, 요청의 httpMethod가 일치하면 해당 메서드 실행(httpRequest 전달)
+                if (methodInfo != null && httpMethod.equals(httpMethodInfo)) {
+                    return (HttpResponse) methodInfo.invoke(o, httpRequest);
+                }
+
+                // 매핑 정보와 일치하지 않으면 예외 발생
+                throw new CustomException(CustomErrorType.INVALID_REQUEST_METHOD);
+            } catch (InvocationTargetException e) {
+                Throwable targetException = e.getTargetException();
+                if (targetException instanceof CustomException) {
+                    throw (CustomException) targetException;
+                }
+            } catch (NoSuchMethodException | InstantiationException | IllegalAccessException e) {
+                logger.error(e.getMessage());
+                throw new CustomException(CustomErrorType.SERVER_ERROR);
             }
-
-            // 매핑 정보와 일치하지 않으면 예외 발생
-            throw new CustomException(CustomErrorType.INVALID_REQUEST_METHOD);
         }
 
         // 정적 페이지 처리
-        return getStaticPage(httpRequest);
-    }
-
-    private HttpResponse getStaticPage(HttpRequest httpRequest) {
-        String filePath = getFilePath(URLUtils.getPath(httpRequest.getURL()));
-        File file = new File(filePath);
-
-        try (FileInputStream fileIn = new FileInputStream(file)) {
-            byte[] body = new byte[(int) file.length()];
-            int readLen = fileIn.read(body);
-            MIMEType mimeType = MIMEType.getMimeType(URLUtils.getExtension(filePath));
-            return HttpResponse
-                    .ok(httpRequest.getVersion())
-                    .contentLength(readLen)
-                    .contentType(mimeType)
-                    .body(body);
-        } catch (IOException e) {
-            throw new CustomException(CustomErrorType.SERVER_ERROR, e);
-        }
-    }
-
-    private String getFilePath(String path) {
-        // 파일 요청이면 그대로 리턴
-        if (!URLUtils.getExtension(path).isEmpty()) {
-            return STATIC_SOURCE_PATH.concat(path);
-        }
-
-        // 해당 url 매핑이 있는지 확인
-        String requestPath = STATIC_MAPPING.keySet().stream()
-                .filter(key -> key.equals(path))
-                .findAny()
-                .orElse("default");
-
-        if (requestPath.equals("default")) {
-            throw new CustomException(CustomErrorType.PATH_NOT_FOUND);
-        }
-
-        return STATIC_SOURCE_PATH.concat(path)
-                .concat("/")    // 구분자
-                .concat(STATIC_MAPPING.get(requestPath));
+        return staticPageRequestManager.getStaticPage(httpRequest);
     }
 
 }
